refactor(sdk): change sdk package structure
