+++ packages/sdk/ts/keys/index.ts
@@ -1,0 +1,2 @@
+export { getAllOnChainVks, compareVks, extractAllVks } from "./verifyingKeys";
+export type { IGetAllVksArgs, IMaciVerifyingKeys, IExtractAllVksArgs, IMaciVks } from "./types";
