+++ packages/sdk/ts/poll/index.ts
@@ -1,0 +1,9 @@
+export { getPoll, getPollParams } from "./poll";
+export type {
+  IGetPollArgs,
+  IGetPollData,
+  IGetPollParamsArgs,
+  IPollParams,
+  IPollJoiningInputs,
+  IPollJoinedInputs,
+} from "./types";
