+++ packages/sdk/ts/tally/index.ts
@@ -1,0 +1,9 @@
+export { generateTallyCommitments } from "./commitments";
+export { verify } from "./verification";
+export type {
+  ITallyData,
+  IVerifyArgs,
+  IGenerateTallyCommitmentsArgs,
+  ITallyCommitments,
+  ITallyVotesInputs,
+} from "./types";
