+++ packages/sdk/ts/trees/index.ts
@@ -1,0 +1,2 @@
+export { generateSignUpTree } from "./stateTree";
+export type { IGenerateSignUpTreeArgs, IGenerateSignUpTree } from "./types";
