+++ packages/sdk/ts/user/index.ts
@@ -1,0 +1,2 @@
+export { isUserRegistered, signup, isJoinedUser } from "./user";
+export type { IJoinedUserArgs, IIsRegisteredUser, IIsJoinedUser, ISignupArgs, IRegisteredUserArgs } from "./types";
