+++ packages/sdk/ts/utils/index.ts
@@ -1,36 +1,7 @@
-export { contractExists, currentBlockTimestamp } from "./contracts";
-export type {
-  TallyData,
-  VerifyArgs,
-  IGetPollArgs,
-  IGetPollData,
-  IGenerateTallyCommitmentsArgs,
-  IGetPollParamsArgs,
-  ITallyCommitments,
-  IPollParams,
-  IRegisteredUserArgs,
-  IParseSignupEventsArgs,
-  ISignupData,
-  ISignupArgs,
-  IJoinedUserArgs,
-  IParsePollJoinEventsArgs,
-  IIsRegisteredUser,
-  IIsJoinedUser,
-  IMaciVerifyingKeys,
-  GetAllVksArgs,
-  ICircuitParams,
-  IExtractAllVksArgs,
-  IMaciVks,
-  IGenSignUpTree,
-  IGenSignUpTreeArgs,
-  IPollJoinedInputs,
-  IPollJoiningInputs,
-  IProcessMessagesInputs,
-  ISnarkJSVerificationKey,
-  ITallyVotesInputs,
-  IVote,
-} from "./types";
-export { verifyPerVOSpentVoiceCredits, verifyTallyResults } from "./verifiers";
 export { BLOCKS_STEP } from "./constants";
-export { parsePollJoinEvents, parseSignupEvents } from "./user";
+export { contractExists, currentBlockTimestamp } from "./contracts";
+export { validateParams } from "./params";
+export { calculateBinaryTreeDepthFromMaxLeaves, calculateQuinTreeDepthFromMaxLeaves } from "./trees";
 export { isArm } from "./utils";
+
+export type { ICircuitParams, ISnarkJSVerificationKey } from "./types";
