+++ packages/sdk/ts/utils/utils.ts
@@ -1,5 +1,3 @@
-import { SNARK_FIELD_SIZE } from "maci-crypto";
-
 import fs from "fs";
 import os from "os";
 
@@ -30,10 +28,3 @@ export const sleep = async (ms: number): Promise<void> => {
     setTimeout(resolve, ms);
   });
 };
-
-/**
- * Run both format check and size check on a salt value
- * @param salt the salt to validate
- * @returns whether it is valid or not
- */
-export const validateSalt = (salt: bigint): boolean => salt < SNARK_FIELD_SIZE;
