+++ packages/sdk/ts/vote/utils.ts
@@ -1,0 +1,8 @@
+import { SNARK_FIELD_SIZE } from "maci-crypto";
+
+/**
+ * Run both format check and size check on a salt value
+ * @param salt the salt to validate
+ * @returns whether it is valid or not
+ */
+export const validateSalt = (salt: bigint): boolean => salt < SNARK_FIELD_SIZE;
